CS311: Homework 1 - Class for each question.

+++ CS311/homework1/homework1.py
@@ -9,8 +9,10 @@ This file holds the collected functions for questions 1-4 of homework
 import os
 import sys
 
+from operator import mul
 from os import path
-from getopt import getopt
+from getopt import getopt, GetoptError
+
 
 class Options(object):
     """
@@ -20,35 +22,13 @@ class Options(object):
     course = None
     term = None
 
-class Main(object):
+
+class CreateDirs(object):
     """
-    Holds answers/functions for all the questions.
+    Question 1
     """
 
-    def usage(self):
-        header = ('Usage: {0} [OPTIONS]...\nRun question for '
-                  'assignment1\n\nMandatory arguments are '
-                  'mandatory.'.format(__file__))
-        arguments = [
-            ('-h', '', 'Show this text'),
-            ('-v', '', 'Enable verbose output'),
-            ('-c,', '--course', 'The course being taught '
-                               '(requires: -t, or --term)'),
-            ('-t,', '--term', 'The term of courses '
-                             '(requires: -c, or --course'),
-        ]
-        
-        usage = [header]
-        usage.extend(["  {0} {1}\t{2}".format(a1, a2, desc) for
-            a1,a2,desc in arguments])
-        return "\n".join(usage)
-
-    def error(self, error_msg):
-        print "ERROR:",error_msg,"\n"
-        print self.usage()
-        sys.exit(1)
-
-    def question1(self, term, course):
+    def create_dirs(self, term, course):
         """
         Creates a directory `./<term>/<course>/` that contains the
         subdirectories: 
@@ -56,8 +36,6 @@ class Main(object):
         and the symlinks:
             handin -> /usr/local/classes/eecs/<term>/<course>/handin
             website -> /usr/local/classes/eecs/<term>/<course>/public_html
-
-        Fulfills requirements for question 1.
 
         :arg term: The term of the school year
         :arg course: Name of a course. Generally in the form:
@@ -105,8 +83,14 @@ class Main(object):
                 elif Options.verbose:
                     print "Directory: {0} already exists.".format(new_dir)
             if path.exists(symlink_base_path):
-                os.symlink(handin_path, 'handin')
-                os.symlink(website_path, 'website')
+                h = path.join(course_path, 'handin')
+                w = path.join(course_path, 'website')
+                if path.exists(h) and not path.exists(handin_path):
+                    os.symlink(handin_path, 'handin')
+                elif path.exists(w) and not path.exists(website_path):
+                    os.symlink(website_path, 'website')
+                elif Options.verbose:
+                        print "Symlinks already exists."
                 if Options.verbose:
                     print "Created symlinks: handin, website."
             elif not os.access(symlink_base_path, os.W_OK):
@@ -120,26 +104,157 @@ class Main(object):
                             "{0}".format(symlink_base_path))
 
         if Options.verbose:
-            print "Finished creating directories for {0}/{1}".format(term, course)
+            print ("Finished creating directories for "
+                   "{0}/{1}".format(term, course))
+
+
+class AlphabetFun(object):
+    """
+    Question 3
+
+    Using names.txt, a 46K text file containing over five-thousand first
+    names, begin by sorting it into alphabetical order. Then working out
+    the alphabetical value for each name, multiply this value by its
+    alphabetical position in the list to obtain a name score.
+
+    For example, when the list is sorted into alphabetical order, COLIN,
+    which is worth 3 + 15 + 12 + 9 + 14 = 53, is the 938th name in the
+    list. So, COLIN would obtain a score of 938 x 53 = 49714.
+
+    What is the total of all the name scores in the file?
+
+    See the posted examples for how to work with a file. For this,
+    A = 1, B = 2, etc. 
+    """
+    ns = open('names.txt', 'r').read().strip('"').split('","')
+    ns.sort()
+
+    @staticmethod
+    def alpha_position(name, names):
+        """
+        1-offset position in name in a list of names
+        """
+        return names.index(name)+1
+
+
+    @staticmethod
+    def alpha_value(name):
+        """
+        Convert a name to its alphabetical value.
+
+        Example COLIN would convert to 53, because::
+        
+            COLIN = 3 + 15 + 12 + 9 + 14 = 53
+
+        :param name: A name to be converted.
+        :type str:
+        """
+        return sum(ord(i)-ord('@') for i in name)
+
+    @classmethod
+    def alpha_total(self, names):
+        """
+        The total score of a list of names.
+        """
+        return sum(self.alpha_position(name, names) * self.alpha_value(name) for name in names)
+
+
+class GreatestProduct(object):
+    """
+    Question 2
+
+    Finds the greatest product of five consecutive digits in the
+    1000-digit number:
+
+          73167176531330624919225119674426574742355349194934
+          96983520312774506326239578318016984801869478851843
+          85861560789112949495459501737958331952853208805511
+          12540698747158523863050715693290963295227443043557
+          66896648950445244523161731856403098711121722383113
+          62229893423380308135336276614282806444486645238749
+          30358907296290491560440772390713810515859307960866
+          70172427121883998797908792274921901699720888093776
+          65727333001053367881220235421809751254540594752243
+          52584907711670556013604839586446706324415722155397
+          53697817977846174064955149290862569321978468622482
+          83972241375657056057490261407972968652414535100474
+          82166370484403199890008895243450658541227588666881
+          16427171479924442928230863465674813919123162824586
+          17866458359124566529476545682848912883142607690042
+          24219022671055626321111109370544217506941658960408
+          07198403850962455444362981230987879927244284909188
+          84580156166097919133875499200524063689912560717606
+          05886116467109405077541002256983155200055935729725
+          71636269561882670428252483600823257530420752963450
+    """
+    number = "".join(open('number.txt', 'r').read().split('\n'))
+
+    @classmethod
+    def greatest_product(self):
+        """
+        Find the greatest product of a sliding window of 5 ints across a
+        1000 digit number.
+         
+         - `for i ...` iterate over all the numbers, up to the 996th digit.
+         - `max( ...` get the max from the list of products
+         - `reduce(mul, ...` get the product of the list
+         - `map(int, ...)` convert the list of strings to a list of ints.
+
+        Note: this is horribly implicit, and therefore not `pythonic` at all.
+        """
+        number = self.number
+        return max(reduce(mul, map(int, number[i:i+5])) for i in xrange(len(number)-4))
+
+class TriangleNumbers(object):
+    """
+    Question 4
+
+    The nth term of the sequence of triangle numbers is given by:
+
+        tn = (1/2)n (n+1)
+
+    so the first ten triangle numbers are:
+
+        1, 3, 6, 10, 15, 21, 28, 36, 45, 55, ...
+
+    By converting each letter in a word to a number corresponding to its
+    alphabetical position and adding these values we form a word value.
+    For example, the word value for SKY is 19 + 11 + 25 = 55 = t10. If
+    the word value is a triangle number then we shall call the word a
+    triangle word.
+
+    Using words.txt, a 16K text file containing nearly two-thousand
+    common English words, how many are triangle words? 
+    """
+    #words = open('words.txt', 'r')
+
+
+class Main(object):
+    """
+    Holds answers/functions for all the questions.
+    """
 
     def run(self):
         """
         Parses the arguments and options passed to the script.
-        Allows for both long and short forms of 'term' and 'course' when
-        running the script.
-
-        :return type: tuple
-        :return: args and options
         """
-        opts, args = getopt(sys.argv[1:], 'c:t:hv', ['--course', '--term'])
+        try:
+            opts, args = getopt(sys.argv[1:], 'c:t:hv', ['--course', '--term'])
+        except GetoptError as err:
+            print err
+            sys.exit(2)
+            #self.error("Option not recognized: {0}".format(opt))
+
+        if len(opts) == 0:
+            self.error("Please pass in at least one option.")
 
         for opt, arg in opts:
             if opt in ('-v'):
                 Options.verbose = True
-            if opt in ('-h'):
+            elif opt in ('-h'):
                 print self.usage()
                 sys.exit()
-            if opt in ('-c', '--course'):
+            elif opt in ('-c', '--course'):
                 Options.course = arg
             elif opt in ('-t', '--term'):
                 Options.term = arg
@@ -147,71 +262,47 @@ class Main(object):
         if Options.term and Options.course:
             self.question1(Options.term, Options.course)       
             sys.exit()
+        elif Options.term or Options.course:
+            self.error("Provided term or course by itself.")
         else:
             self.error("Not enough options given")
 
+    def usage(self):
+        header = ('Usage: {0} [OPTIONS]...\nRun question for '
+                  'assignment1\n\nMandatory arguments are '
+                  'mandatory.'.format(__file__))
+        arguments = [
+            ('-h', '', 'Show this text'),
+            ('-v', '', 'Enable verbose output'),
+            ('-c,', '--course', 'The course being taught '
+                               '(requires: -t, or --term)'),
+            ('-t,', '--term', 'The term of courses '
+                             '(requires: -c, or --course'),
+        ]
+        
+        usage = [header]
+        usage.extend(["  {0} {1}\t{2}".format(a1, a2, desc) for
+            a1,a2,desc in arguments])
+        return "\n".join(usage)
 
-"""
-Question 2
+    def error(self, error_msg):
+        print "ERROR:",error_msg,"\n"
+        #print self.usage()
+        sys.exit(1)
 
-Finds the greatest product of five consecutive digits in the 1000-digit number:
+    def question1(self, term, course):
+        CreateDirs().create_dirs(term, course)
 
-      73167176531330624919225119674426574742355349194934
-      96983520312774506326239578318016984801869478851843
-      85861560789112949495459501737958331952853208805511
-      12540698747158523863050715693290963295227443043557
-      66896648950445244523161731856403098711121722383113
-      62229893423380308135336276614282806444486645238749
-      30358907296290491560440772390713810515859307960866
-      70172427121883998797908792274921901699720888093776
-      65727333001053367881220235421809751254540594752243
-      52584907711670556013604839586446706324415722155397
-      53697817977846174064955149290862569321978468622482
-      83972241375657056057490261407972968652414535100474
-      82166370484403199890008895243450658541227588666881
-      16427171479924442928230863465674813919123162824586
-      17866458359124566529476545682848912883142607690042
-      24219022671055626321111109370544217506941658960408
-      07198403850962455444362981230987879927244284909188
-      84580156166097919133875499200524063689912560717606
-      05886116467109405077541002256983155200055935729725
-      71636269561882670428252483600823257530420752963450
-"""
+    def question2(self):
+        print GreatestProduct().greatest_product()
 
-"""
-Question 3
+    def question3(self):
+        print AlphabetFun().alpha_total(AlphabetFun.ns)
 
-Using names.txt, a 46K text file containing over five-thousand first names,
-begin by sorting it into alphabetical order. Then working out the alphabetical
-value for each name, multiply this value by its alphabetical position in the
-list to obtain a name score.
-
-For example, when the list is sorted into alphabetical order, COLIN, which is
-worth 3 + 15 + 12 + 9 + 14 = 53, is the 938th name in the list. So, COLIN would
-obtain a score of 938 x 53 = 49714.
-
-What is the total of all the name scores in the file?
-
-See the posted examples for how to work with a file. For this, A = 1, B = 2,
-etc. 
-"""
-
-"""
-Question 4
-
-The nth term of the sequence of triangle numbers is given by, tn = (1/2)n (n+1);
-so the first ten triangle numbers are:
-
-1, 3, 6, 10, 15, 21, 28, 36, 45, 55, ...
-
-By converting each letter in a word to a number corresponding to its
-alphabetical position and adding these values we form a word value. For
-example, the word value for SKY is 19 + 11 + 25 = 55 = t10. If the word value
-is a triangle number then we shall call the word a triangle word.
-
-Using words.txt, a 16K text file containing nearly two-thousand common English
-words, how many are triangle words? 
-"""
 
 if __name__ == "__main__":
+    #Main().question1('fa13','cs421')
+    #Main().question2()
+    #Main().question3()
     Main().run()
+    #Main().greatest_product()
